Create hist.php
this is the new history page. Both tables work. bugs are fixed, sizing is fixed. Still not able to grab the specific Tool_ID value from a specific row. but for the meantime,  I made it so the table always grabs from WHERE Tool_ID=1. So every button is grabbing from the same thing.